# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FitSphere API",
        "description": "API Documentation for FitSphere",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/gyms": {
            "get": {
                "tags": [
                    "Gyms"
                ],
                "summary": "List all gyms",
                "operationId": "5917b5dba336b46002602c9a494922af",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter by gym name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/gyms/{id}": {
            "get": {
                "tags": [
                    "Gyms"
                ],
                "summary": "Get gym details",
                "operationId": "cd1f00d34bee7532692328029901894d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Gym ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Gym not found"
                    }
                }
            }
        },
...
        "/products": {
            "get": {
                "tags": [
                    "Store - Products"
                ],
                "summary": "List all products with filtering and sorting",
                "operationId": "c1ff6f862214e1896c59cfe0491ce0e8",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category name (e.g. Supplements, Gear)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Minimum price",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Maximum price",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort option (price_asc, price_desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "price_asc",
                                "price_desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Store - Products"
                ],
                "summary": "Get product details",
                "operationId": "b46f4aa73232d0368eee698a082342a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/trainers": {
            "get": {
                "tags": [
                    "Trainers"
                ],
                "summary": "List all trainers with advanced filtering and sorting",
                "operationId": "ee552e84cfffce7d4ca2567030d8a29a",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter by trainer name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "speciality",
                        "in": "query",
                        "description": "Filter by specialization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Minimum price per month",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Maximum price per month",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "min_rating",
                        "in": "query",
                        "description": "Minimum average rating (e.g. 4.0)",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort by option (top_rated, price_asc, price_desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "top_rated",
                                "price_asc",
                                "price_desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/trainers/{id}": {
            "get": {
                "tags": [
                    "Trainers"
                ],
                "summary": "Get trainer profile details",
                "operationId": "22ea11cdf2e876885bef0353061e7b3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Trainer ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Trainer not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}